[v1.0.0] Profile and optimize the formatters
Added a note regarding the recent GitHub action failures because of
the `getpass.getuser` breaking on windows. Link to the issue is added
for a quick reference. Any open PRs due to this should be closed by the
maintainers.

Deprecate use of regex for extracting path length from `LOGGER_MSG_FMT`
over a limit constant. This reduced the function calls exponentially.

Remove `__all__` as the module does not intent to export the formatters.

Updated and reorganized all the docstrings to keep updated with latest
modifications. This fixes the typos and the internal references.
Fix note regarding the docstrings and module's intent.

Add support for singleton design pattern in the formatters. Fix class
docstrings and typos. Attribute names have been changed to make more
sense and easy to understand. Formatter class now uses full pathname
instead of abspath of the current directory.
Fixed `pathname` hack by setting a new `caller` attribute to LogRecord
object.

Added support for positional-only arguments in `StreamHandler`. This
makes the intent of coloring more clear and enforces users to specify
what should be colorized.