Upstream new package structure
This package structure is inspired from `pytest` and follows a similar
pattern. This structure provides both implict and an explicit way of
accessing the APIs.

A new `py.typed` is also added, however, not sure how that is useful.

Finally, I've dropped the idea of having a dedicated config module in
favor of dynamic monkey patching behavior.